[JENKINS-72240] Add top level license declarations
https://issues.jenkins.io/browse/JENKINS-72240 notes that the plugin
does not declare the license in its pom.xml or in a LICENSE file. Many
of the source files include an MIT license header.  No other license
is visible in the source files, except MIT license, so it seems safe to
declare the plugin as using the MIT license.

Dates in the copyright notice were extracted from the git commits and
the license headings added to source files.